RAiD-270 - created auth and constants folders, moved keycloak to auth
