+++ raid-agency-app/src/entities/organisation/forms/organisation-details-form/OrganisationDetailsForm.tsx
@@ -1,6 +1,6 @@
 import { OrganisationLookupButton } from "@/containers/organisation-lookup/OrganisationLookupButton";
 import { OrganisationLookupDialog } from "@/containers/organisation-lookup/OrganisationLookupDialog";
-import { TextInputField } from "@/fields/TextInputField";
+import { TextInputField } from "@/components/fields/TextInputField";
 import { IndeterminateCheckBox } from "@mui/icons-material";
 import { Grid, IconButton, Stack, Tooltip, Typography } from "@mui/material";
 import { useQuery } from "@tanstack/react-query";
