+++ raid-agency-app/src/entities/subject/forms/subject-details-form/SubjectDetailsForm.tsx
@@ -1,4 +1,4 @@
-import { TextSelectField } from "@/fields/TextSelectField";
+import { TextSelectField } from "@/components/fields/TextSelectField";
 import subjectMapping from "@/mapping/data/subject-mapping.json";
 import { IndeterminateCheckBox } from "@mui/icons-material";
 import { Grid, IconButton, Stack, Tooltip, Typography } from "@mui/material";
