+++ raid-agency-app/src/services/auth-service.ts
@@ -1,5 +1,5 @@
 // src/services/authService.ts
-import { keycloakInstance } from "@/keycloak";
+import { keycloakInstance } from "@/auth/keycloak";
 
 // Define roles as readonly const to ensure type safety
 export const REALM_ROLES = {
